codeql: widen two multiplications before the cast; exclude world-writable
cpp/integer-multiplication-cast-to-long fired on two int/float products
evaluated in the narrow type before widening. Both operands are bounded here
(config array sizes; a frame count times ms-per-frame), so neither overflows —
but making the widening explicit is the correct idiom and keeps the query live
to catch a genuine overflow later, so fix rather than exclude.

cpp/world-writable-file-creation is the opposite case: 15 instances of the same
portable std::fopen(path, "wb") idiom, whose created-file mode comes from the
process umask. The fix the query wants is POSIX-only (open with an explicit
mode) and would break the MSVC job, and file permissions are the caller's
concern, not a weakness in a library writing a file it was asked to write.
Exclude it in the CodeQL config alongside path-injection, with the rationale.

+++ src/kokoro_modules.cpp
@@ -1316,7 +1316,7 @@ void Generator::load_from(const stf::File& f, const KokoroConfig& cfg) {
     }
 
     // ─── resblocks ────────────────────────────────────────────────────────
-    resblocks.resize(num_upsamples * num_kernels);
+    resblocks.resize(static_cast<std::size_t>(num_upsamples) * num_kernels);
     for (int i = 0; i < num_upsamples; ++i) {
         const int C_cur = init_C / (1 << (i + 1));
         for (int j = 0; j < num_kernels; ++j) {
